Added migration support to update
Added descriptions to the DataRecord type.

I think that the migrate function still needs more clarity, as the name parameter of the function is unclear.  It could mean the record name, the property name, or the database name.  I would interpret property name as the others are known upon calling.  However it is unclear in the writing.