Removed migration support and updated design
Migration is an important feature which can be added later, but is of lower priority.  Migration support depends on :load()'s implementation.

To get the data from DataRecords, :data() is no longer necessary.  Developers can now just index the record to retrieve the external data.  This may feature may not make the design.

:fetch() clashed w/ :data() and was unclear, so it was renamed to fetchRecord.

Instead of loading a record into the database, we want a clear representation that a record will be created and data can be loaded onto the record using the :load(data) function.

Added GetProperty and withProperty.  SetProperty has been updated to not return a Database.