Adjusted the specification to exclude migration
Migration aspects involve restructuring and enforcing the structure to be of a certain form.  These aspects are not directly correlated to creation aspects.  The mistake of assuming migration was necessary to bake-in was due to the relation between the creation of the structure and restructuring.  Migration support should be easy to add to Creator, I may even create another module doing such.