Separated migrator and replicator examples
Modified the implementation to be capable of supporting extensions in the database builder / new record phase.

Fixed migrator, it is not complete since encode/decode are not defined.

Replicator was not modified

A compressor example was provided, which has yet to be complete, but it restructures the data to be more optimal for serialization.  Most likely, not the most optimal, since that approach seems to be too costly to implement.