Remove attributes from design + add more extension support
The first iteration didn't have support for attributes, it was added later out of the need to allow for the data to have some metadata.  It can easily be re-added via an extension.

Extension support was also adjusted to allow for extenders to chose between cloning a table and referencing it.  Also added a way to avoid certain elements that should solely be in the database, such as the version, from being added to records.  Also added a way to hide properties from the database itself, mainly used when building the database and you need to have certain internal properties.  There may be more needs.