Removal of variadic/dynamic support
Support for records whose element can be adjusted (added/removed) can be extended.  ALSO there is already a roundabout approach to attach records using a small Database.

RecordBuilder by proxy can now be removed, same w/ the builder method in DataRecord.

I am considering naming Database -> Template since it is not a database.  It is moreso a template that a record is assumed to follow, that can be extended to MUST follow.

Also removed hide/avoid since they should be thought about more prior to truly adopting the concepts.  Avoid is easy to implement via extension, as shown in the migration example, so it will not be reconsidered.  For hide, I don't have an example planned that involves, but it may be necessary.