Database -> Struct and add withValue/value functions to replace .Value
Struct sounds better, since it allows for multiple meanings.  One of which is by C lang users, which is partially correct.  It is moreso a loosely defined struct, since extenders can redefine the internal layout.

value() was added since there is no guarantee that .Value will produce the value.  In fact, it may be empty, BUT the extender wants it to point to the default value.

I may do record(name) instead of forcing extenders to modify the __index metamethod when records are no longer storing as `a.Foo`.  Uncertain.