Adjusted the examples to be closer to the intended use
Creator now solely uses extension aspects to implement its default form, which includes Value, Parent, and Name, which are universally useful for extenders and users alike.

To provide extenders with better approach to modify record how records are gotten, record(name) was added to the Record type.

Compressor has yet to be completed.