Added RecordWhenTable?
It may later be removed, support for extension via these sort of options is rather finiky and does not involve the overall design.

The temporary goal is to make compressor work, then ensure that extensions mesh well together.  If the design needs to be adjusted to better accommodate this, I would rather know before I finalize it.