Add CloneIfTable to internal properties
May later be removed.  I noticed that if .Value is a table, then sending a reference would not be good.  Should "clone" have been used, then Value could only be a table.  So this should be the best of both worlds.  And if users truly want the reference, they can set __extend.Value = true.