Adjusted replicator example
When the replicator function was called, it was impossible to tell which record was changed when handling multiple records.

withValue no longer needs to be updated in the example w/ all.  Same for FullName, it is no longer necessary.

Replicator no longer has withReplicator due to the above issue.  In its place, onChange was added to Record.  Replicator was drastically simplified due to this adjustment and all direct performance concerns have been dealt w/.

The purpose of onChange in example w/ all is unclear as of writing, it may be extended to support changes to attributes (withAttribute) and changes to records (withRecord).