Add mode parameter to Replicator
Also changed unknown back to any for defaults, since any & A = A.

Mode can either be "Attribute" or "Record" in this case, where it is nil when the value itself has changed.  In my own usage case, I will probably use replicator behind the scenes and not provide it for users.

If we could remove StructFinder that would be nice, compressor uses it solely for getting the Id.  While we could just remove the Id aspect, it is a major part of saving output size.