Add attribute compression to example w/ all
Also added attribute and record replication.

The prior mentioned removing structfinder, but that would entail two new methods to Struct.  reformatRecord and deformatRecord where we either duplicate the table and reformat the table to output a smaller json or take the reformatted table and reconstruct it into the original.  The cost on memory is too great, especially in cases that need to quickly serialize changes.