Added records to Creator
Records was removed earlier, but after careful consideration it has been re-added.  Especially since users don't currently have an adequate way to type the Record objects for their own usage case.

Records was modified in compressor to pass back the actual name.  Records does NOT pass the actual table used to store the records, not for safety but to ensure consistency among extensions.

Added inverserelation, which is used to make records much faster.